refactor: merge openai responses vision proxy test.
